Added Working Template folder
MPLab is buggy asf and has a gajillion project setup requisites. Just use this one already pre-setup folder to complete the stuff and either copy files or make each commit a new project or do the projects on a commit timeline for safety and also to be able to look back.